+++ src/frontend/src/components/YourTrips/CommuteGroup.tsx
@@ -1,15 +1,18 @@
 import { useEffect, useState } from "react";
 
+import selectNeighborhoodZipcodeMap from "reducers/neighborhoods/selectors/selectNeighborhoodZipcodeMap";
+import type { Destination } from "reducers/userProfile/types";
 import { useAppSelector } from "store/store";
+
 import { Accordion } from "components/base/Accordion/Accordion";
 import { AccordionItem } from "components/base/Accordion/AccordionItem";
 import Range from "components/base/Range/Range";
-import { yourTripsStyles } from "./YourTrips.styles";
-import selectNeighborhoodZipcodeMap from "reducers/neighborhoods/selectors/selectNeighborhoodZipcodeMap";
 import { createGoogleDirectionsURL } from "libs/getLinkURLs";
-import type { TripType } from "./types";
-import type { Destination } from "reducers/userProfile/types";
+import { MAX_TRAVEL_TIME } from "src/constants";
+import { NonAccordionCommuteButton } from "./NonAccordionCommuteButton";
 import TripMap from "./TripMap";
+import type { TripType } from "./types";
+import { yourTripsStyles } from "./YourTrips.styles";
 
 const CommuteGroup = ({
     trips,
@@ -25,6 +28,7 @@ const CommuteGroup = ({
     });
     const [selectedMapDestination, setSelectedMapDestination] =
         useState<Destination | null>(null);
+    const [selectedDestIsTooFar, setSelectedDestIsTooFar] = useState(false);
     const allNeighborhoodCommutes = useAppSelector(
         selectNeighborhoodZipcodeMap
     );
@@ -36,6 +40,11 @@ const CommuteGroup = ({
 
     useEffect(() => {
         setSelectedMapDestination(initialTrip.destination);
+        const minCommuteTime =
+            allNeighborhoodCommutes[initialTrip.neighborhoodZipcode]?.commutes[
+                initialTrip.destination.location.label
+            ].commuteMin;
+        setSelectedDestIsTooFar(minCommuteTime > MAX_TRAVEL_TIME);
     }, [initialTrip.destination]);
 
     const getCommuteMin = (trip: TripType) =>
@@ -53,7 +62,7 @@ const CommuteGroup = ({
         if (selection) {
             // TODO will need to adjust for compare page
             setSelectedMapDestination(trips[parseInt(selection)].destination);
-            if (!isTransit) {
+            if (!isTransit || selectedDestIsTooFar) {
                 // open link
                 const directionsURL = createGoogleDirectionsURL(
                     trips[parseInt(selection)].neighborhoodZipcode,
@@ -68,37 +77,65 @@ const CommuteGroup = ({
 
     return (
         <div className={styles.commuteGroupWrapper()}>
-            <Accordion
-                defaultExpandedKeys={["0"]}
-                expandedItemCallback={onExpandItem}
-                className={styles.commuteGroupItem()}
-            >
-                {trips.map((trip, index) => (
-                    <AccordionItem
-                        id={index.toString()}
-                        key={index}
-                        title={trip.title}
-                        subtitle={trip.subtitle}
-                        titleContentRight={
-                            <Range
-                                start={getCommuteMin(trip)}
-                                end={getCommuteMax(trip)}
-                            />
-                        }
-                        overridePanelOpen={!isTransit}
-                        isMobile={isMobile}
-                    >
-                        {isMobile && (
-                            <TripMap
-                                start={trip.destination}
-                                end={initialTrip.neighborhoodZipcode}
-                                isTransit={isTransit}
-                            />
-                        )}
-                    </AccordionItem>
-                ))}
-            </Accordion>
-            {!isMobile && selectedMapDestination && (
+            {trips?.length === 1 ? (
+                <NonAccordionCommuteButton
+                    id={"0"}
+                    key={0}
+                    title={trips[0].title}
+                    subtitle={trips[0].subtitle}
+                    titleContentRight={
+                        <Range
+                            start={getCommuteMin(trips[0])}
+                            end={getCommuteMax(trips[0])}
+                        />
+                    }
+                    overridePanelOpen={!isTransit || selectedDestIsTooFar}
+                    isMobile={isMobile}
+                    onPress={() => onExpandItem(new Set("0"))}
+                >
+                    {isMobile && (
+                        <TripMap
+                            start={trips[0].destination}
+                            end={initialTrip.neighborhoodZipcode}
+                            isTransit={isTransit}
+                        />
+                    )}
+                </NonAccordionCommuteButton>
+            ) : (
+                <Accordion
+                    defaultExpandedKeys={["0"]}
+                    expandedItemCallback={onExpandItem}
+                    className={styles.commuteGroupItem()}
+                >
+                    {trips.map((trip, index) => (
+                        <AccordionItem
+                            id={index.toString()}
+                            key={index}
+                            title={trip.title}
+                            subtitle={trip.subtitle}
+                            titleContentRight={
+                                <Range
+                                    start={getCommuteMin(trip)}
+                                    end={getCommuteMax(trip)}
+                                />
+                            }
+                            overridePanelOpen={
+                                !isTransit || selectedDestIsTooFar
+                            }
+                            isMobile={isMobile}
+                        >
+                            {isMobile && (
+                                <TripMap
+                                    start={trip.destination}
+                                    end={initialTrip.neighborhoodZipcode}
+                                    isTransit={isTransit}
+                                />
+                            )}
+                        </AccordionItem>
+                    ))}
+                </Accordion>
+            )}
+            {!isMobile && selectedMapDestination && !selectedDestIsTooFar && (
                 <TripMap
                     start={selectedMapDestination}
                     end={initialTrip.neighborhoodZipcode}