+++ src/frontend/src/components/base/Checkbox/Checkbox.styles.ts
@@ -3,16 +3,15 @@ import { tv } from "tailwind-variants";
 export const checkboxStyles = tv({
     slots: {
         base: "flex items-start gap-3",
-        box: "flex h-6 w-6 flex-shrink-0 items-center justify-center rounded-[var(--spacing-3)] border-2 transition",
+        box: "flex flex-shrink-0 items-center justify-center rounded-[var(--spacing-3)] transition",
         labelContainer: "flex flex-col gap-1",
         label: "text-rg font-bold text-gray-900",
         description: "text-sm text-gray-600",
+        footer: "text-sm text-gray-600 flex flex-row gap-2 items-center mt-2",
     },
     variants: {
         isSelected: {
-            true: {
-                box: "border-teal-700 bg-teal-700",
-            },
+            true: {},
             false: {
                 box: "border-gray-400 bg-white",
             },
@@ -29,5 +28,37 @@ export const checkboxStyles = tv({
                 description: "text-gray-400 hover:cursor-not-allowed",
             },
         },
+        variant: {
+            default: {},
+            ghost: {},
+        },
+        size: {
+            default: {
+                box: "h-6 w-6 border-2",
+            },
+            small: {
+                box: "h-[22px] w-[22px] border-1",
+            },
+        },
+    },
+    compoundVariants: [
+        {
+            isSelected: true,
+            variant: "default",
+            class: {
+                box: "border-teal-700 bg-teal-700 text-white",
+            },
+        },
+        {
+            isSelected: true,
+            variant: "ghost",
+            class: {
+                box: "bg-white border-gray-400 text-gray-600",
+            },
+        },
+    ],
+    defaultVariants: {
+        variant: "default",
+        size: "default",
     },
 });