+++ src/frontend/src/pages/Discover/Desktop.tsx
@@ -1,34 +1,20 @@
 import { UserProfileSubheader } from "components/UserProfileSubheader";
-import { Place } from "src/enums";
 import discoverStyles from "./Discover.styles";
-import Neighborhoods from "./Neighborhoods";
 import Map from "./Map/Map";
-import NeighborhoodDetail from "pages/NeighborhoodDetail";
+import ModalsWrapper from "./ModalsWrapper";
+import Neighborhoods from "./Neighborhoods";
 
-const DiscoverDesktop = ({
-    isDetailModalOpen,
-    setIsDetailModalOpen,
-}: {
-    isDetailModalOpen: boolean;
-    setIsDetailModalOpen: (b: boolean) => void;
-}) => {
+const DiscoverDesktop = () => {
     const { root, sidebar, headerContainer } = discoverStyles({
         isMobile: false,
     });
 
     return (
         <div className={root()}>
-            <NeighborhoodDetail
-                modalOpen={isDetailModalOpen}
-                modalOpenChangeCallback={setIsDetailModalOpen}
-            />
+            <ModalsWrapper />
             <div className={sidebar()}>
                 <div className={headerContainer()}>
-                    <UserProfileSubheader
-                        size="medium"
-                        mode="transit"
-                        place={Place.Work}
-                    />
+                    <UserProfileSubheader size="medium" />
                 </div>
                 <Neighborhoods mobile={false} />
             </div>