+++ src/frontend/src/pages/Discover/Mobile.tsx
@@ -2,19 +2,12 @@ import { useEffect } from "react";
 import { useSearchParams } from "react-router";
 
 import { UserProfileSubheaderMobile } from "components/UserProfileSubheader";
-import { Place } from "src/enums";
 import discoverStyles from "./Discover.styles";
-import Neighborhoods from "./Neighborhoods";
 import Map from "./Map/Map";
-import NeighborhoodDetail from "../NeighborhoodDetail";
+import ModalsWrapper from "./ModalsWrapper";
+import Neighborhoods from "./Neighborhoods";
 
-const DiscoverMobile = ({
-    isDetailModalOpen,
-    setIsDetailModalOpen,
-}: {
-    isDetailModalOpen: boolean;
-    setIsDetailModalOpen: (b: boolean) => void;
-}) => {
+const DiscoverMobile = () => {
     const [routerParams, setRouterParams] = useSearchParams();
     const display = routerParams.get("display");
 
@@ -37,16 +30,10 @@ const DiscoverMobile = ({
 
     return display ? (
         <div className={root()}>
-            <NeighborhoodDetail
-                modalOpen={isDetailModalOpen}
-                modalOpenChangeCallback={setIsDetailModalOpen}
-                isMobile
-            />
+            <ModalsWrapper isMobile />
             <div className={headerContainer()}>
                 <UserProfileSubheaderMobile
                     size="small"
-                    mode="transit"
-                    place={Place.Work}
                     display={display}
                     callback={onChangeMode}
                 />