+++ src/frontend/src/pages/Discover/ModalsWrapper.tsx
@@ -1,0 +1,13 @@
+import EditFiltersModal from "src/components/EditFiltersModal";
+import NeighborhoodDetail from "../NeighborhoodDetail";
+import EditTripsModal from "src/components/EditTripsModal";
+
+const ModalsWrapper = ({ isMobile }: { isMobile?: boolean }) => (
+    <>
+        <NeighborhoodDetail isMobile={isMobile} />
+        <EditFiltersModal isMobile={isMobile} />
+        <EditTripsModal />
+    </>
+);
+
+export default ModalsWrapper;