+++ src/frontend/src/reducers/neighborhoods/neighborhoodsSlice.ts
@@ -1,10 +1,14 @@
 import { createSelector, createSlice } from "@reduxjs/toolkit";
+import { NEIGHBORHOOD_REGIONS } from "src/constants";
+import type { RootState } from "store/store";
+import { getNeighborhoodsAndBounds } from "./neighborhoodsThunk";
 import type {
+    FiltersState,
+    NeighborhoodBounds,
+    NeighborhoodProperties,
     NeighborhoodsSliceState,
     RankedNeighborhoodsLists,
 } from "./types";
-import { getNeighborhoodsAndBounds } from "./neighborhoodsThunk";
-import type { RootState } from "store/store";
 
 const initialState: NeighborhoodsSliceState = {
     neighborhoods: null,
@@ -12,6 +16,7 @@ const initialState: NeighborhoodsSliceState = {
     activeNeighborhood: null,
     loading: false,
     error: null,
+    filters: { ecc: false, regions: NEIGHBORHOOD_REGIONS, textSearch: null },
     rankCalculating: false,
     rankedNeighborhoodsLists: {
         topTen: [],
@@ -46,6 +51,12 @@ export const neighborhoodSlice = createSlice({
             state.rankedNeighborhoodsLists = lists;
             state.rankCalculating = false;
         },
+        setNeighborhoodFilters: (
+            state,
+            { payload: filters }: { payload: FiltersState }
+        ) => {
+            state.filters = filters;
+        },
     },
     extraReducers: builder => {
         builder
@@ -70,6 +81,7 @@ export const {
     setActiveNeighborhood,
     setRankCalculating,
     setRankedNeighborhoodLists,
+    setNeighborhoodFilters,
 } = neighborhoodSlice.actions;
 
 export const selectRankedNeighborhoodsLists = (state: RootState) =>
@@ -86,6 +98,50 @@ export const selectActiveNeighborhoodBounds = (state: RootState) =>
     state.neighborhoods.neighborhoodBounds?.features.find(
         b => b.properties.zipcode === state.neighborhoods.activeNeighborhood
     );
+export const selectNeighborhoodPropsByZipcode = (state: RootState) =>
+    state.neighborhoods.neighborhoods?.features.reduce(
+        (neighborhoodPropsMap, neighborhood) => ({
+            ...neighborhoodPropsMap,
+            [neighborhood.properties.zipcode]: neighborhood.properties,
+        }),
+        {} as Record<string, NeighborhoodProperties>
+    );
+export const selectAreFiltersApplied = (state: RootState) => {
+    const searchTerm = state.neighborhoods.filters.textSearch?.trim();
+    return (
+        state.neighborhoods.filters.ecc ||
+        state.neighborhoods.filters.regions?.length <
+            NEIGHBORHOOD_REGIONS.length ||
+        !!searchTerm
+    );
+};
+export const selectNeighborhoodFilters = (state: RootState) =>
+    state.neighborhoods.filters;
+
+export const selectFilterableNeighborhoodBounds = createSelector(
+    [
+        selectRankedNeighborhoodsLists,
+        selectAreFiltersApplied,
+        (state: RootState) => state.neighborhoods.neighborhoodBounds,
+        // Filters state unused in fn, but passed in to pick up changes
+        (state: RootState) => state.neighborhoods.filters,
+    ],
+    (rankedNeighborhoodsLists, isFiltered, neighborhoodBounds, _) => {
+        const { topTen, recommended, tooFar } = rankedNeighborhoodsLists;
+        if (isFiltered) {
+            return {
+                type: "FeatureCollection",
+                features: neighborhoodBounds?.features.filter(f =>
+                    // Filtered rankedNeighborhoodsLists groups
+                    [...topTen, ...recommended, ...tooFar].includes(
+                        f.properties.zipcode
+                    )
+                ),
+            } as NeighborhoodBounds;
+        }
+        return neighborhoodBounds;
+    }
+);
 
 export const selectNeighborhoodNameByZipcode = createSelector(
     [(state: RootState) => state.neighborhoods.neighborhoods],