+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add top 10 tour [#730](https://github.com/azavea/echo-locator/pull/730)
 - Add neighborhood detail preview cards to Discover page [#741](https://github.com/azavea/echo-locator/pull/741)
 - Add base Trips Wizard [#740](https://github.com/azavea/echo-locator/pull/740)
+- Add neighborhood list display filters [#738](https://github.com/azavea/echo-locator/pull/738)
 
 ### Changed